Add files via upload
This document has the EDA section of the  report as well in addition to what we saw in the meeting. Please review and let me know in case of any more review comments..

Based on what we discussed:
1) I will make update to the Project Objective
2) Add the Mobile net details of the model and mention other models planned..

Please do share any other inputs to be incorporated so that we can finalize this tomorrow.